requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==36.0.2
faiss-cpu==1.11.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
